backend/requirements.txt
# Flask Backend Requirements for PMO Portfolio

# Core Flask
Flask==2.3.3
Flask-CORS==4.0.0

# Environment management
python-dotenv==1.0.0

# Azure Databricks connectivity
databricks-sql-connector==2.9.3

# Data processing
pandas==2.1.1
numpy==1.24.3

# HTTP requests
requests==2.31.0

# JSON handling
jsonschema==4.19.1

# Logging and monitoring
structlog==23.1.0

# Date/time handling
python-dateutil==2.8.2

# Security
cryptography==41.0.4

# Development dependencies
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production server
gunicorn==21.2.0
